+++ apps/block_scout_web/lib/block_scout_web/api_router.ex
@@ -46,7 +46,7 @@ defmodule BlockScoutWeb.ApiRouter do
     plug(RateLimit)
   end
 
-  alias BlockScoutWeb.Account.Api.V1.{AuthenticateController, TagsController, UserController}
+  alias BlockScoutWeb.Account.Api.V1.{AuthenticateController, EmailController, TagsController, UserController}
   alias BlockScoutWeb.API.V2
 
   scope "/account/v1", as: :account_v1 do
@@ -57,6 +57,10 @@ defmodule BlockScoutWeb.ApiRouter do
     post("/authenticate", AuthenticateController, :authenticate_post)
 
     get("/get_csrf", UserController, :get_csrf)
+
+    scope "/email" do
+      get("/resend", EmailController, :resend_email)
+    end
 
     scope "/user" do
       get("/info", UserController, :info)