+++ apps/block_scout_web/lib/block_scout_web/controllers/address_transaction_controller.ex
@@ -216,72 +216,57 @@ defmodule BlockScoutWeb.AddressTransactionController do
     end
   end
 
+  defp items_csv(
+         conn,
+         %{
+           "address_id" => address_hash_string,
+           "from_period" => from_period,
+           "to_period" => to_period
+         },
+         csv_export_module
+       )
+       when is_binary(address_hash_string) do
+    with {:ok, address_hash} <- Chain.string_to_address_hash(address_hash_string),
+         {:ok, address} <- Chain.hash_to_address(address_hash),
+         true <- Application.get_env(:block_scout_web, :recaptcha)[:is_disabled] do
+      address
+      |> csv_export_module.export(from_period, to_period)
+      |> Enum.reduce_while(put_resp_params(conn), fn chunk, conn ->
+        case Conn.chunk(conn, chunk) do
+          {:ok, conn} ->
+            {:cont, conn}
+
+          {:error, :closed} ->
+            {:halt, conn}
+        end
+      end)
+    else
+      :error ->
+        unprocessable_entity(conn)
+
+      {:error, :not_found} ->
+        not_found(conn)
+
+      false ->
+        not_found(conn)
+    end
+  end
+
   defp items_csv(conn, _, _), do: not_found(conn)
 
   def token_transfers_csv(conn, params) do
-    items_csv(
-      conn,
-      %{
-        "address_id" => params["address_id"],
-        "from_period" => params["from_period"],
-        "to_period" => params["to_period"],
-        "recaptcha_response" => params["recaptcha_response"]
-      },
-      AddressTokenTransferCsvExporter
-    )
+    items_csv(conn, params, AddressTokenTransferCsvExporter)
   end
 
-  def transactions_csv(conn, %{
-        "address_id" => address_hash_string,
-        "from_period" => from_period,
-        "to_period" => to_period,
-        "recaptcha_response" => recaptcha_response
-      }) do
-    items_csv(
-      conn,
-      %{
-        "address_id" => address_hash_string,
-        "from_period" => from_period,
-        "to_period" => to_period,
-        "recaptcha_response" => recaptcha_response
-      },
-      AddressTransactionCsvExporter
-    )
+  def transactions_csv(conn, params) do
+    items_csv(conn, params, AddressTransactionCsvExporter)
   end
 
-  def internal_transactions_csv(conn, %{
-        "address_id" => address_hash_string,
-        "from_period" => from_period,
-        "to_period" => to_period,
-        "recaptcha_response" => recaptcha_response
-      }) do
-    items_csv(
-      conn,
-      %{
-        "address_id" => address_hash_string,
-        "from_period" => from_period,
-        "to_period" => to_period,
-        "recaptcha_response" => recaptcha_response
-      },
-      AddressInternalTransactionCsvExporter
-    )
+  def internal_transactions_csv(conn, params) do
+    items_csv(conn, params, AddressInternalTransactionCsvExporter)
   end
 
-  def logs_csv(conn, %{
-        "address_id" => address_hash_string,
-        "from_period" => from_period,
-        "to_period" => to_period,
-        "recaptcha_response" => recaptcha_response
-      }) do
-    items_csv(
-      conn,
-      %{
-        "address_id" => address_hash_string,
-        "from_period" => from_period,
-        "to_period" => to_period,
-        "recaptcha_response" => recaptcha_response
-      },
-      AddressLogCsvExporter
-    )
+  def logs_csv(conn, params) do
+    items_csv(conn, params, AddressLogCsvExporter)
   end
 end